+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 - Updated deps: `std`, `parcel_css`, `cliffy`.
+- The list of available plugins on `lume init`.
 
 ## [1.11.1] - 2022-09-02
 ### Fixed